Reinstate root files
Many of the files in the root had been turning into symlinks. This
reverses that change. Instead, we will copy the files before publishing
the `create` package.
