+++ .gitignore
@@ -1,1 +1,7 @@
-lib/create/skel/.gitignore
+*~
+*.log
+coverage/
+node_modules/
+packages/*/node_modules/
+package-lock.json
+storybook-static/
