--- .jest
@@ -1,1 +1,0 @@
-lib/create/skel/.jest
