--- lib/create/skel/.gitignore
@@ -1,1 +1,0 @@
-gitignore
