--- lib/create/skel/tsconfig.build.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "composite": true,
-    "declaration": true,
-    "module": "ES2020",
-    "sourceMap": false,
-    "target": "ES2019"
-  }
-}
