--- lib/create/skel/tsconfig.nodejs.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig.build.json",
-  "compilerOptions": {
-    "module": "commonjs",
-    "removeComments": true,
-  }
-}
