--- lib/create/skel/tsconfig.webpack.json
@@ -1,19 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "ES2020",
-    "target": "ES2020",
-    "types": [
-      "webpack-env"
-    ]
-  },
-  "include": [
-    "*/*/src/**/*"
-  ],
-  "exclude": [
-    "dist",
-    "node_modules",
-    "*/*/dist",
-    "*/*/node_modules"
-  ]
-}
