+++ tsconfig.nodejs.json
@@ -1,1 +1,7 @@
-lib/create/skel/tsconfig.nodejs.json
+{
+  "extends": "./tsconfig.build.json",
+  "compilerOptions": {
+    "module": "commonjs",
+    "removeComments": true,
+  }
+}
